# specification/docs/api-design/openapi.yaml
openapi: 3.0.0
info:
  title: サークル練習参加抽選システム API
  description: 練習日への投票・重み付き抽選・結果確認を提供する REST API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: アカウント登録（プロフィール込み）
      description: |
        登録直後はメール未確認状態。登録したメールアドレス宛に6桁の確認コードを送信し、
        `POST /auth/verify` で確認が完了するまでログインできない (REQ-001.1 / D-011)。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 登録成功（メール未確認）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: メールアドレスが登録済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify:
    post:
      tags: [auth]
      summary: メールアドレス確認（6桁コード）
      description: |
        確認コードの有効期限は15分。入力を5回失敗するとコードが無効化され、
        再送するまで確認できない (REQ-001.5 / D-012)。

        確認に成功するとアクセストークンを返す。ログイン画面へ戻さず、
        そのままログイン状態にすること (D-014)。

        エラーは登録済みメールアドレスを判別できないよう、以下をすべて 400 で返す。

        | code | 状況 |
        |------|------|
        | `INVALID_CODE` | コード不一致、未登録のメール、確認済み |
        | `CODE_EXPIRED` | 有効期限切れ |
        | `CODE_LOCKED` | 失敗回数が上限に到達 |
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        # POST の 200 は原則未定義だが、確認完了と同時にログイン状態にするため例外的に定義 (D-014)
        '200':
          description: 確認完了（アクセストークンを発行し、そのままログイン状態になる）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/verify/resend:
    post:
      tags: [auth]
      summary: 確認コードの再送
      description: |
        新しいコードを発行し、有効期限と失敗回数をリセットする。古いコードは無効になる (REQ-001.6)。
        メールアドレスの存在を漏らさないため、未登録・確認済みの場合も 204 を返す。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '204':
          description: 再送受付（未登録・確認済みでも 204 を返す）
  /auth/login:
    post:
      tags: [auth]
      summary: ログイン（JWT発行）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        # POST の 200 は原則未定義だが、トークン返却が必須のため例外的に定義
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: メール未確認（`EMAIL_NOT_VERIFIED`）。確認コードの入力を促すこと
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/password-reset/request:
    post:
      tags: [auth]
      summary: パスワードリセットメール送信
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: 送信受付（存在しないメールでも 204 を返す）
  /auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: 新パスワード設定
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        '204':
          description: 再設定成功
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /practice-months/{pmId}/vote-summary:
    get:
      tags: [lottery]
      summary: 抽選前の投票状況と学年別枠（代表のみ）
      description: |
        REQ-005.5 / D-015。練習日ごとの学年別投票者数、代表が設定済みの参加人数枠、
        システムの提案値を返す。抽選前の調整画面が使う。

        `quotas_ready` が false の練習日があると抽選は実行できない。
      parameters:
        - $ref: '#/components/parameters/PmId'
      responses:
        '200':
          description: 投票状況と枠
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteSummary'
        '403':
          $ref: '#/components/responses/Forbidden'
  /practice-months/{pmId}/quotas:
    put:
      tags: [lottery]
      summary: 学年別の参加人数枠を保存（代表のみ）
      description: |
        REQ-005.3 / D-015。練習日ごとに1〜3年の人数をまとめて保存する。
        **学年別人数の合計が定員と一致しない場合は 400** を返す。
      parameters:
        - $ref: '#/components/parameters/PmId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotaUpdateRequest'
      responses:
        '204':
          description: 保存成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /practice-months/{pmId}/lottery:
    post:
      tags: [lottery]
      summary: 抽選の一括実行（代表のみ）
      description: REQ-005。再実行時は既存実行が無効化される（REQ-005.12、confirm_rerun 必須）
      parameters:
        - $ref: '#/components/parameters/PmId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LotteryExecuteRequest'
      responses:
        '201':
          description: 抽選完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LotteryExecution'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 投票期間中は実行不可 / 再実行に confirm_rerun が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /practice-months/{pmId}/executions:
    get:
      tags: [lottery]
      summary: 抽選実行履歴（代表のみ）
      parameters:
        - $ref: '#/components/parameters/PmId'
      responses:
        '200':
          description: 実行履歴
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LotteryExecution'
        '403':
          $ref: '#/components/responses/Forbidden'
  /practice-months/{pmId}/results/me:
    get:
      tags: [lottery]
      summary: 自分の当選結果（公開後のみ）
      parameters:
        - $ref: '#/components/parameters/PmId'
      responses:
        '200':
          description: 自分の参加日一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyResults'
        '404':
          description: 未公開または存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /practice-months/{pmId}/results:
    get:
      tags: [lottery]
      summary: 全結果の取得（代表のみ。練習日別・メンバー別）
      parameters:
        - $ref: '#/components/parameters/PmId'
      responses:
        '200':
          description: 全結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullResults'
        '403':
          $ref: '#/components/responses/Forbidden'
  /practices/{practiceId}/assignments:
    post:
      tags: [lottery]
      summary: 参加者の手動追加（代表のみ・微調整）
      description: REQ-006.3。定員超過時は 201 で警告フラグを返す
      parameters:
        - $ref: '#/components/parameters/PracticeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreateRequest'
      responses:
        '201':
          description: 追加成功（capacity_exceeded で定員超過を警告）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentCreated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 既に割当済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /assignments/{assignmentId}:
    delete:
      tags: [lottery]
      summary: 参加者の手動削除（代表のみ・微調整）
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: string, example: asg_01H0000000000000000000000 }
      responses:
        '204':
          description: 削除成功
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /practice-months/{pmId}/publish:
    post:
      tags: [lottery]
      summary: 結果の公開（代表のみ）
      description: REQ-006.4。公開後にメンバーが /results/me を閲覧できる。月次実績 (monthly_member_results) もこの時点で確定
      parameters:
        - $ref: '#/components/parameters/PmId'
      responses:
        '204':
          description: 公開成功
        '409':
          description: 抽選未実行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PmId:
      name: pmId
      in: path
      required: true
      schema: { type: string, example: pmn_01H0000000000000000000000 }
    PracticeId:
      name: practiceId
      in: path
      required: true
      schema: { type: string, example: prc_01H0000000000000000000000 }
  responses:
    BadRequest:
      description: 入力不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未認証・トークン無効
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 権限不足・性別スコープ違反
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Gender:
      type: string
      enum: [male, female]
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  reason: { type: string }
    RegisterRequest:
      type: object
      required: [email, password, name, address, phone_number, grade, gender, faculty_department, student_number, is_manager]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
        address: { type: string }
        phone_number: { type: string }
        grade: { type: integer, enum: [1, 2, 3] }
        gender: { $ref: '#/components/schemas/Gender' }
        faculty_department: { type: string }
        student_number: { type: string }
        is_manager: { type: boolean }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, description: 秒 }
    VerifyEmailRequest:
      type: object
      required: [email, code]
      properties:
        email: { type: string, format: email }
        code:
          type: string
          pattern: '^\d{6}$'
          example: '482915'
          description: メールで届いた6桁の確認コード
    ResendVerificationRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetConfirm:
      type: object
      required: [token, new_password]
      properties:
        token: { type: string }
        new_password: { type: string, minLength: 8 }
    UserProfile:
      type: object
      properties:
        id: { type: string, example: usr_01H0000000000000000000000 }
        email: { type: string, format: email }
        name: { type: string }
        address: { type: string }
        phone_number: { type: string }
        grade: { type: integer, enum: [1, 2, 3] }
        gender: { $ref: '#/components/schemas/Gender' }
        faculty_department: { type: string }
        student_number: { type: string }
        is_manager: { type: boolean }
        role: { type: string, enum: [member, representative] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Practice:
      type: object
      properties:
        id: { type: string, example: prc_01H0000000000000000000000 }
        practice_date: { type: string, format: date }
        starts_at: { type: string, example: '18:00' }
        ends_at: { type: string, example: '21:00' }
        location: { type: string, example: 第一体育館 }
        capacity: { type: integer, example: 30 }
        vote_count: { type: integer, description: 現在の投票数（一覧表示用） }
...
    VoteSummary:
      type: object
      description: 抽選前の調整画面に出す、その月の投票状況と枠 (D-015)
      properties:
        practice_month_id: { type: string }
        quotas_ready:
          type: boolean
          description: 全練習日の全学年に枠が設定済みか（false なら抽選できない）
        practices:
          type: array
          items:
            $ref: '#/components/schemas/PracticeVoteSummary'
    PracticeVoteSummary:
      type: object
      properties:
        practice_id: { type: string }
        practice_date: { type: string, format: date }
        starts_at: { type: string, example: '18:00' }
        ends_at: { type: string, example: '21:30' }
        location: { type: string }
        capacity: { type: integer }
        grades:
          type: array
          description: 3年 → 1年の順
          items:
            type: object
            properties:
              grade: { type: integer, enum: [1, 2, 3] }
              voters: { type: integer, description: その日に投票した人数 }
              quota:
                type: integer
                nullable: true
                description: 代表が設定した参加人数枠。未設定なら null
              suggested_quota:
                type: integer
                description: システムの提案値（REQ-005.5.1）
    QuotaUpdateRequest:
      type: object
      description: 学年別の参加人数枠をまとめて保存する (D-015)
      required: [practices]
      properties:
        practices:
          type: array
          items:
            type: object
            required: [practice_id, grades]
            properties:
              practice_id: { type: string }
              grades:
                type: array
                description: 1〜3年すべてを指定する。合計は定員と一致させること
                items:
                  type: object
                  required: [grade, quota]
                  properties:
                    grade: { type: integer, enum: [1, 2, 3] }
                    quota: { type: integer, minimum: 0 }
    LotteryExecuteRequest:
      type: object
      properties:
        seed: { type: integer, nullable: true, description: 省略時はサーバーが生成（再現実行用） }
        confirm_rerun: { type: boolean, default: false, description: 再実行時は true 必須（REQ-005.12） }
    LotteryExecution:
      type: object
      properties:
        id: { type: string, example: lot_01H0000000000000000000000 }
        practice_month_id: { type: string }
        executed_by: { type: string }
        random_seed: { type: integer }
        grade2_ratio:
          type: number
          nullable: true
          description: D-015 で廃止。実行時の学年別枠は settings_snapshot.quotas に記録する
        is_active: { type: boolean }
        warnings:
          type: array
          description: 3年定員超過などの警告（REQ-005.3）
          items: { type: string }
        created_at: { type: string, format: date-time }
    MyResults:
      type: object
      properties:
        practice_month_id: { type: string }
        assignments:
          type: array
          items:
            type: object
            properties:
              practice: { $ref: '#/components/schemas/Practice' }
    FullResults:
      type: object
      properties:
        by_practice:
          type: array
          items:
            type: object
            properties:
              practice: { $ref: '#/components/schemas/Practice' }
              participants:
                type: array
                items:
                  $ref: '#/components/schemas/Participant'
        by_member:
          type: array
          items:
            type: object
            properties:
              user_id: { type: string }
              name: { type: string }
              grade: { type: integer }
              is_manager: { type: boolean }
              votes_count: { type: integer }
              wins_count: { type: integer }
              practice_ids:
                type: array
                items: { type: string }
    Participant:
      type: object
      properties:
        assignment_id: { type: string }
        user_id: { type: string }
        name: { type: string }
        grade: { type: integer }
        is_manager: { type: boolean }
        assigned_via:
          type: string
          enum: [manager, grade3, guaranteed, distribution, overflow, manual]
    AssignmentCreateRequest:
      type: object
      required: [user_id]
      properties:
        user_id: { type: string }
    AssignmentCreated:
      type: object
      properties:
        assignment_id: { type: string }
        capacity_exceeded: { type: boolean, description: 定員超過の警告（REQ-006.3） }
...